数据库一体机:完成bin2sql的数据修复功能

+++ dataRepair/bin2sql/binlog2sqlUtilils.py
@@ -11,23 +11,13 @@ class binlogUtilis:
         # 是不是DML事件
         if isinstance(event, WriteRowsEvent) or isinstance(event, UpdateRowsEvent) or isinstance(event,
                                                                                                  DeleteRowsEvent):
-            return True
+            return 1
+        elif (isinstance(event, QueryEvent) and event.query.lower().startswith("alter table")) \
+                or (isinstance(event, QueryEvent) and event.query.lower().startswith("drop table")) \
+                or (isinstance(event, QueryEvent) and event.query.lower().startswith("create table")):
+            return 2
         else:
-            return False
-
-
-    @classmethod
-    def concat_sql_from_binlog_event(cls, cursor, binlog_event, row=None, e_start_pos=None, flashback=False,
-                                     no_pk=False):
-        if flashback and no_pk:
-            raise ValueError('only one of flashback or no_pk can be True')
-        if not (isinstance(binlog_event, WriteRowsEvent) or isinstance(binlog_event, UpdateRowsEvent)
-                or isinstance(binlog_event, DeleteRowsEvent) or isinstance(binlog_event, QueryEvent)):
-            raise ValueError('binlog_event must be WriteRowsEvent, UpdateRowsEvent, DeleteRowsEvent or QueryEvent')
-
-        if isinstance(binlog_event, WriteRowsEvent) or isinstance(binlog_event, UpdateRowsEvent) \
-                or isinstance(binlog_event, DeleteRowsEvent):
-            cls.generate_sql_pattern(binlog_event)
+            return -1
 
     @classmethod
     def compare_items(cls, items):
@@ -39,42 +29,41 @@ class binlogUtilis:
             return '`%s`=%%s' % k
 
     @classmethod
-    def fix_object(cls, value):
-        """Fixes python objects so that they can be properly inserted into SQL queries"""
-        if isinstance(value, set):
-            value = ','.join(value)
-        else:
-            return value
-
-    @classmethod
-    def confirmSchem(cls, database,tableName, values):
+    def confirmSchem(cls, database, tableName, values):
         if os.path.exists(f"./schemaGroup/{database}/{tableName}.json"):
             with open(f"./schemaGroup/{database}/{tableName}.json", encoding="utf-8") as fp:
                 tableSchema = json.load(fp)
                 if len(values) == len(tableSchema[tableName]["columns"]):
                     # 主键和所有键值
-                    priColumns = {}
-                    corColumns = {tableSchema[tableName]['columns'][i]: values[i] for i in range(len(values))}
+                    primary_key_columns = {}
+                    other_columns = {tableSchema[tableName]['columns'][i]: values[i] for i in range(len(values))}
                     for i in tableSchema[tableName]["primaryKey"]:
-                        priColumns[i] = corColumns[i]
-                    return corColumns, priColumns
+                        primary_key_columns[i] = other_columns[i]
+                    return other_columns, primary_key_columns
                 else:
                     raise TypeError(f"{tableName}的表结构没有还原成原始的样子")
 
     @classmethod
-    def generate_sql_pattern(cls, binlog_event, replace=False):
+    def generate_sql_pattern(cls, binlog_event, database, table, replace=False):
+        # 判断这个是不是目前监听的表
+        if database and binlog_event.schema not in database:
+            return
+        if table and f"{binlog_event.schema}.{binlog_event.table}" not in table:
+            return
+        # 生成sql语句
         if isinstance(binlog_event, DeleteRowsEvent):
             columns, priColumns = cls.confirmSchem(binlog_event.schema, binlog_event.table,
                                                    list(binlog_event.rows[0]['values'].values()))
+            values = priColumns.values()
             template = 'DELETE FROM `{0}`.`{1}` WHERE {2} LIMIT 1;'.format(
                 binlog_event.schema, binlog_event.table,
                 ' AND '.join(map(cls.compare_items, list(priColumns.items())))
             )
-            values = priColumns.values()
-            print(111, template, values)
+            return [template, list(values)]
         elif isinstance(binlog_event, WriteRowsEvent):
             # 获取字段对应
-            columns, priColumns = cls.confirmSchem(binlog_event.schema, binlog_event.table,list(binlog_event.rows[0]['values'].values()))
+            columns, priColumns = cls.confirmSchem(binlog_event.schema, binlog_event.table,
+                                                   list(binlog_event.rows[0]['values'].values()))
             if not replace:
                 values = columns.values()
                 template = 'INSERT INTO `{0}`.`{1}`({2}) VALUES ({3});'.format(
@@ -82,6 +71,7 @@ class binlogUtilis:
                     ', '.join(map(lambda i: f"`{i}`", columns.keys())),
                     ', '.join(['%s'] * len(values))
                 )
+                return [template, list(values)]
             else:
                 values = columns.values()
                 template = 'REPLACE INTO `{0}`.`{1}`({2}) values({3})'.format(
@@ -89,10 +79,10 @@ class binlogUtilis:
                     ', '.join(map(lambda i: f"`{i}`", columns.keys())),
                     ', '.join(['%s'] * len(values))
                 )
-            print(111, template, values)
+                return [template, list(values)]
 
         elif isinstance(binlog_event, UpdateRowsEvent):
-            columns, priColumns = cls.confirmSchem(binlog_event.schema,binlog_event.table,
+            columns, priColumns = cls.confirmSchem(binlog_event.schema, binlog_event.table,
                                                    list(binlog_event.rows[0]['after_values'].values()))
             if not replace:
                 template = 'UPDATE `{0}`.`{1}` SET {2} WHERE {3} LIMIT 1;'.format(
@@ -101,6 +91,7 @@ class binlogUtilis:
                     ' AND '.join(map(cls.compare_items, list(priColumns.items()))))
                 values = list(columns.values())
                 values.extend(priColumns.values())
+                return [template, values]
             else:
                 values = columns.values()
                 template = 'REPLACE INTO `{0}`.`{1}`({2}) values({3})'.format(
@@ -108,5 +99,4 @@ class binlogUtilis:
                     ', '.join(map(lambda i: f"`{i}`", columns.keys())),
                     ', '.join(['%s'] * len(values))
                 )
-
-            print(111,template, values)
+                return template, values
